+++ server/gamenode/gameserver.js
@@ -7,6 +7,7 @@ const https = require('https');
 const fs = require('fs');
 
 const config = require('./nodeconfig.js');
+const { detectBinary } = require('../util');
 const logger = require('../log.js');
 const ZmqSocket = require('./zmqsocket.js');
 const Game = require('../game/game.js');
@@ -105,7 +106,7 @@ class GameServer {
         let debugData = {};
 
         if(e.message.includes('Maximum call stack')) {
-            debugData.badSerializaton = this.detectBinary(gameState);
+            debugData.badSerializaton = detectBinary(gameState);
         } else {
             debugData.game = gameState;
             debugData.game.players = undefined;
@@ -123,32 +124,6 @@ class GameServer {
         if(game) {
             game.addMessage('A Server error has occured processing your game state, apologies.  Your game may now be in an inconsistent state, or you may be able to continue.  The error has been logged.');
         }
-    }
-
-    detectBinary(state, path = '', results = []) {
-        const allowedTypes = ['Array', 'Boolean', 'Date', 'Number', 'Object', 'String'];
-
-        if(!state) {
-            return results;
-        }
-
-        let type = state.constructor.name;
-
-        if(!allowedTypes.includes(type)) {
-            results.push({ path: path, type: type });
-        }
-
-        if(type === 'Object') {
-            for(let key in state) {
-                this.detectBinary(state[key], `${path}.${key}`, results);
-            }
-        } else if(type === 'Array') {
-            for(let i = 0; i < state.length; ++i) {
-                this.detectBinary(state[i], `${path}[${i}]`, results);
-            }
-        }
-
-        return results;
     }
 
     clearStaleFinishedGames() {
